Component looks real again
* messages-app, message-event component templates
* messages-app subscribes to WS for new messages
* declare global JS in typescript as ADVOCATE const
* compute events_empty when updating events
* comment out a lot of existing JS (like jquery)
