+++ app/assets/components/messages/src/index.ts
@@ -2,8 +2,9 @@ import App from './main';
 import { ComponentManager, setPropertyDidChange } from '@glimmer/component';
 import initializeCustomElements from '@glimmer/web-component';
 
+declare const ADVOCATE;
+
 const app = new App();
-const containerElement = document.getElementById('messages-app');
 
 setPropertyDidChange(() => {
   app.scheduleRerender();
@@ -15,7 +16,10 @@ app.registerInitializer({
   }
 });
 
-app.renderComponent('messages-app', containerElement, null);
-
-app.boot();
-initializeCustomElements(app, ['messages-app']);
+// GLOBAL from application.js
+ADVOCATE.document_ready(() => {
+  const containerElement = document.getElementById('messages-app'); // div id=messages-app
+  app.renderComponent('messages-app', containerElement, null);
+  app.boot();
+  initializeCustomElements(app, ['messages-app']); // <messages-app>
+});
