+++ app/assets/components/messages/src/ui/components/message-event/component.ts
@@ -1,0 +1,4 @@
+import Component from '@glimmer/component';
+
+export default class MessageEvent extends Component {
+};
