+++ app/assets/components/messages/src/ui/components/message-event/template.hbs
@@ -1,0 +1,19 @@
+<div class="message">
+{{#if @event.message_id}}
+  <div class="text-bubble incoming">
+    {{ @event.content }}
+  </div>
+  <p class="incoming-msg-from">
+    {{ @event.created_at }} {{!-- .strftime("%m/%d/%Y %l:%M %p") --}}
+  </p>
+{{else}}
+  <div class="text-bubble outgoing">
+    {{ @event.content }}
+  </div>
+  <p class="incoming-msg-to">
+    {{!-- current_user.first_name --}}
+    <span class="provider-name"></span>
+    {{ @event.created_at }} {{!-- .strftime("%m/%d/%Y %l:%M %p") --}}
+  </p>
+{{/if}}
+</div>
