+++ app/assets/stylesheets/application.scss
@@ -1,2 +1,2 @@
 @import 'advocate';
-@import 'selectize.min';
+// @import 'selectize.min';
