+++ app/views/layouts/application.html.erb
@@ -12,6 +12,7 @@
   <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro" rel="stylesheet">
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.15.1/moment.min.js"></script>
+  <%= javascript_include_tag 'application' %>
 </head>
 <body>
 
@@ -20,7 +21,5 @@
   <div class="page-container">
     <%= yield %>
   </div>
-
-  <%= javascript_include_tag 'application' %>
 </body>
 </html>
